perf_hooks: fix loop delay resolution validation
Fixes: https://github.com/nodejs/node/issues/38160

PR-URL: https://github.com/nodejs/node/pull/38166

+++ src/node_perf.h
@@ -166,7 +166,7 @@ class ELDHistogram : public IntervalHistogram {
   ELDHistogram(
       Environment* env,
       v8::Local<v8::Object> wrap,
-      int32_t interval);
+      int64_t interval);
 
   void OnInterval() override;
